requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
sqlalchemy>=2.0.35
apscheduler>=3.10.4
polygon-api-client>=0.1.9
yfinance>=0.2.40
pandas-market-calendars>=4.3.1
pytz>=2024.1
jinja2>=3.1.3
python-multipart>=0.0.6
python-dotenv>=1.0.0
requests>=2.31.0
